+++ Plugin.cs
@@ -2,24 +2,42 @@
 using Steamworks;
 using HarmonyLib;
 using System.Reflection;
+using BepInEx.Logging;
 
 namespace GetYourOwnPortal;
 
 [BepInPlugin(Plugin.GUID, Plugin.NAME, Plugin.VERSION)]
 public class Plugin : BaseUnityPlugin
 {
+    static ManualLogSource Logger = new ManualLogSource(GUID);
+
     // Plugin metadata
     public const string NAME = "GetYourOwnPortal";
     public const string AUTHOR = "Kevver";
     public const string GUID = $"{AUTHOR}.{NAME}";
-    public const string VERSION = "1.0.0";
+    public const string VERSION = "1.0.1";
 
     // HarmonyLib handle
     static readonly Harmony harmony = new(NAME);
 
+    public static ServerSync.ConfigSync configSync = new ServerSync.ConfigSync(GUID)
+    {
+        DisplayName = NAME,
+        CurrentVersion = VERSION,
+        MinimumRequiredVersion = VERSION
+    };
+
+    public static bool PlayerIsAdmin()
+    {
+        bool isAdmin = configSync.IsAdmin || (ZNet.instance && ZNet.instance.IsServer());
+        Logger.LogInfo($"IsAdmin: {isAdmin}");
+        return isAdmin;
+    }
+
     private void Awake()
     {
         // Setup loggers
+        BepInEx.Logging.Logger.Sources.Add(Plugin.Logger);
         BepInEx.Logging.Logger.Sources.Add(Portal.Create.Logger);
         BepInEx.Logging.Logger.Sources.Add(Portal.Edit.Logger);
         BepInEx.Logging.Logger.Sources.Add(Portal.Destroy.Logger);