+++ CHANGELOG.md
@@ -1,3 +1,9 @@
 ## Changelog
+### 1.0.1
+- Added ServerSync
+- Fixed admin check using ServerSync.ConfigSync
+- Fixed portal destruction check
+- Fixed privilege check for disowning a portal
+
 ### 1.0.0
 - Initial release